#56 refactor: S3Service 위치 변경(domain.recruitment.service -> global.util)
