+++ src/main/java/com/boaz/backend/domain/recruitment/service/RecruitmentService.java
@@ -22,6 +22,7 @@ import com.boaz.backend.domain.recruitment.repository.SubscriptionRepository;
 import com.boaz.backend.global.common.enums.Track;
 import com.boaz.backend.global.exception.CustomException;
 import com.boaz.backend.global.exception.ErrorCode;
+import com.boaz.backend.global.util.S3Service;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
